backend/requirements.txt
Flask==3.0.3
networkx==3.3
forceatlas2-python==1.1
python-dotenv==1.0.1
gunicorn==23.0.0